CAP: Implement CAP/CL/ADV/BV-04-C
- Add wid_421
- Move announcement function to generic cap file
- Test requires available contexts to be 0 in both adv and PACS
